requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
psycopg2-binary==2.8.6
